Refactor db calls for buffet
