+++ util/buffetHelpers.js
@@ -1,5 +1,7 @@
 const Buffet = require("../db/models/buffet");
 
+const { getBuffet } = require("../db/actions/buffet");
+
 /**
  * This method searches for the most recent buffet menu.
  * If no buffet menu is found, unset buffet is set to true and returns empty buffet items.
@@ -13,91 +15,18 @@ const Buffet = require("../db/models/buffet");
  */
 async function getLatestBuffet() {
   try {
-    let currentBuffet = await Buffet.findOne({}).sort("-updatedAt");
-    let unsetBuffet = false;
+    const currentBuffet = await getBuffet();
 
-    // Check if the buffet exists
-    if (!currentBuffet) {
-      unsetBuffet = true;
-      currentBuffet = {
-        menuItems: [],
-        extraItems: [],
-      };
-    } else {
-      // If the buffet exists, populate the menu for the buffet
-      await currentBuffet.populate("menuItems").execPopulate();
-
-      // If no menu exists for the buffet, intialize it to empty array.
-      if (
-        !currentBuffet.menuItems ||
-        !currentBuffet.menuItems instanceof Array
-      ) {
-        currentBuffet.menuItems = [];
-      }
-
-      // If no extra items for the buffet exists, initialize it to empty array
-      if (
-        !currentBuffet.extraItems ||
-        !currentBuffet.extraItems instanceof Array
-      ) {
-        currentBuffet.extraItems = [];
-      }
-
-      // If no menu item or extra item exists, then buffet can be considered as unset
-      if (
-        currentBuffet.extraItems.length == 0 &&
-        currentBuffet.menuItems.length == 0
-      ) {
-        unsetBuffet = true;
-      }
-    }
+    // If no menu item or extra item exists, then buffet can be considered as unset
+    let unsetBuffet =
+      currentBuffet.extraItems.length == 0 &&
+      currentBuffet.menuItems.length == 0;
 
     return { unsetBuffet, currentBuffet };
   } catch (err) {
     console.log(err);
     throw "Failed to get the latest buffet";
   }
-}
-
-/**
- * This method takes a buffet object to be added to the database.
- * Check models/Buffet.js for the schema of Buffet
- * The caller must handle the exceptions thrown.
- *
- * @param {obj} buffet new buffet menu to be added to the database
- * @returns {obj} newly created buffet menu document
- * @public
- */
-async function createBuffet(buffet) {
-  Buffet.create(buffet, (err, buffet) => {
-    if (err) {
-      console.log(err);
-      throw "Failed to create a new buffet";
-    } else {
-      return buffet;
-    }
-  });
-}
-
-/**
- * This method updates the buffet based on the buffet id and new buffet menu
- * passed to it.
- * The caller must handle the exceptions thrown.
- *
- * @param {mongoose.ObjectId | string} buffetId id of the buffet
- * @param {obj} newBuffet updated buffet menu
- * @returns {obj} updated buffet document
- * @public
- */
-async function updateBuffet(buffetId, newBuffet) {
-  Buffet.findByIdAndUpdate(buffetId, newBuffet, (err, buffet) => {
-    if (err) {
-      console.log(err);
-      throw "Failed to update the buffet";
-    } else {
-      return buffet;
-    }
-  });
 }
 
 /**
@@ -117,12 +46,7 @@ function getEmptyBuffetItem() {
   };
 }
 
-/**
- * This module contains methods required to access database to perform operations for Buffet
- */
 module.exports = {
   getLatestBuffet,
-  createBuffet,
-  updateBuffet,
   getEmptyBuffetItem,
 };
